Replace digital POT with PWM for reference voltage
In an effort to simplify the circuit board, the digital POT has been replaced
by the use of PWM for the reference voltage. The calibration algorithm has also
been updated obviating the need to use the Arduino's analog inputs to determine
the appropriate reference voltage setting. Instead, the algorithm now
determines an appropriate reference voltage by measuring the dutycycle of the
signal out of the voltage comparator.